IMP - Refactor unit name reference files
